FIX: Broken locales in context menu
