+++ javascripts/discourse/components/spreadsheet-editor.js
@@ -12,7 +12,6 @@ import { popupAjaxError } from "discourse/lib/ajax-error";
 import I18n from "I18n";
 import { schedule } from "@ember/runloop";
 import { tracked } from "@glimmer/tracking";
-import { localeMapping } from "../../discourse-table-builder/lib/locale-mapping";
 
 export default class SpreadsheetEditor extends Component {
   @tracked showEditReason = false;
@@ -175,7 +174,7 @@ export default class SpreadsheetEditor extends Component {
       defaultColAlign: "left",
       wordWrap: true,
       csvFileName: exportFileName,
-      text: localeMapping,
+      text: this.localeMapping(),
       ...opts,
     });
   }
@@ -239,4 +238,49 @@ export default class SpreadsheetEditor extends Component {
 
     return arrayToTable(table, headers);
   }
+
+  localeMapping() {
+    return {
+      noRecordsFound: prefixedLocale("no_records_found"),
+      show: prefixedLocale("show"),
+      entries: prefixedLocale("entries"),
+      insertANewColumnBefore: prefixedLocale("context_menu.col.before"),
+      insertANewColumnAfter: prefixedLocale("context_menu.col.after"),
+      deleteSelectedColumns: prefixedLocale("context_menu.col.delete"),
+      renameThisColumn: prefixedLocale("context_menu.col.rename"),
+      orderAscending: prefixedLocale("context_menu.order.ascending"),
+      orderDescending: prefixedLocale("context_menu.order.descending"),
+      insertANewRowBefore: prefixedLocale("context_menu.row.before"),
+      insertANewRowAfter: prefixedLocale("context_menu.row.after"),
+      deleteSelectedRows: prefixedLocale("context_menu.row.delete"),
+      copy: prefixedLocale("context_menu.copy"),
+      paste: prefixedLocale("context_menu.paste"),
+      saveAs: prefixedLocale("context_menu.save"),
+      about: prefixedLocale("about"),
+      areYouSureToDeleteTheSelectedRows: prefixedLocale(
+        "prompts.delete_selected_rows"
+      ),
+      areYouSureToDeleteTheSelectedColumns: prefixedLocale(
+        "prompts.delete_selected_cols"
+      ),
+      thisActionWillDestroyAnyExistingMergedCellsAreYouSure: prefixedLocale(
+        "prompts.will_destroy_merged_cells"
+      ),
+      thisActionWillClearYourSearchResultsAreYouSure: prefixedLocale(
+        "prompts.will_clear_search_results"
+      ),
+      thereIsAConflictWithAnotherMergedCell: prefixedLocale(
+        "prompts.conflict_with_merged_cells"
+      ),
+      invalidMergeProperties: prefixedLocale("invalid_merge_props"),
+      cellAlreadyMerged: prefixedLocale("cells_already_merged"),
+      noCellsSelected: prefixedLocale("no_cells_selected"),
+    };
+  }
+}
+
+function prefixedLocale(localeString) {
+  return I18n.t(
+    themePrefix(`discourse_table_builder.spreadsheet.${localeString}`)
+  );
 }
